Make cvxpy an optional dependency (Qiskit/qiskit-aqua#1055)
* Make cvxpy an optional dependency

This commit makes the cvxpy dependency optional because many users have
trouble installing it because on many environments it requires that a
compiler be present to build the package from sdist and there are no
precompiled binary wheels present for many environments. Users of the
svm classifier will have to install cvxpy manually or with the extra
now, but it should remove the barrier for most users.

Related to Qiskit/qiskit-aqua#1053

* Fix lint and requirements list

* Update requirements list

* add cvxpy install to travis, remove from requirements-dev.txt

* Update error message

* Update releasenotes/notes/cvx-optional-e44920f87cff5e8d.yaml

Co-authored-by: Steve Wood <40241007+woodsp-ibm@users.noreply.github.com>

Co-authored-by: Manoel Marques <manoel@us.ibm.com>
Co-authored-by: Steve Wood <40241007+woodsp-ibm@users.noreply.github.com>