Allow input_gradients to be set in constructor of NeuralNetwork (#117)
Summary: This PR introduces a parameter ``input_gradients`` to the constructor of ``NeuralNetwork`` and its subclasses. ``input_gradients`` defaults to False. The constructor of ``TorchConnector`` no longer sets ``input_gradients=True`` by default on the ``NeuralNetwork`` it was instantiated with. This PR closes #90.

Detail and comments: If you are using ``TorchConnector`` and want to compute gradients w.r.t. to inputs, you will have to set ``input_gradients`` explicitly on the ``NeuralNetwork`` you pass to ``TorchConnector``.

Changelog: New Feature

* Reword docstring

* Remove input_gradients parameter from VQC and VQR constructors and instantiate their NN with input_gradients=False

Co-authored-by: jonvet <jonas.vetterle.16@ucl.ac.uk>
Co-authored-by: Anton Dekusar <62334182+adekusar-drl@users.noreply.github.com>