Review batch support in NNClassifier/Regressor (#102)
This is a continuation of the batch support work. There are a few batch related TODOs left in the objective functions used by NeuralNetworkClassifier and NeuralNetworkRegressor. Right now these functions don't make use of the batch support added in the neural network implementation and compute all elements in a for loop. The goal of this this PR is to review these functions, add batch support and review loss function implementations as they don't support batches as well.
This PR closes #63.

Co-authored-by: Steve Wood <40241007+woodsp-ibm@users.noreply.github.com>