Move away from using pylint header plugin (#133)
* Move away from using pylint header plugin

Currently we're relying on a pylint plugin to verify copyright headers
conform to the project guidelines. However this is problematic for a
number of reasons. The first is it has proven to be unstable there have
been several instances of both users and CI failing on header files that
do not have any issues. This is pretty common with pylint in general
since it's very sensitive to differences in environment. Additionally,
it was using a regex to try and match the entire header string. This
provided no practical debugability when it fails (especially when the
failure is unexpected) because we end up with only a match error.
This commit addresses these issues by removing the use of the regex
based pylint plugin and copying and adapting the small script qiskit-terra
uses to verify the header in every qiskit file and provide useful debugging
if there is an issue with a header in a python file.

* Add verify_headers arguments to makefile and tox.ini

Co-authored-by: Manoel Marques <manoel.marques@ibm.com>