Fix forward pass computation cache (#213)
Objective functions cache result of forward computation. Initially this cache was based on the identifier (a call of id()) of the weights array. AQGD re-uses the same array for weights: it updates the values keeping an instance of the array the same. This caused to re-use the same forward pass value across all iteration. The forward pass cache now is based on actual values of weights.

Co-authored-by: Manoel Marques <Manoel.Marques@ibm.com>