Transpile circuit in CircuitQNN for better performance (#243)
Currently every time when CircuitQNN executes a circuit it is transpiled internally by QuantumInstance. Thus, overall fir time of a model based on CircuitQNN is larger than it could be. In this PR we introduce an additional transpilation step in CircuitQNN that is done when a quantum instance is set. The circuit is transpiled and saved for further re-use. So, every time when the circuit is executed it is already transpiled and overall time of the forward pass is reduced.

* add transpilation to CircuitQNN.

* fix spell

* added reno

* fix indentation in reno

* fix spell

* rephrase the warning

* suppress warning

* rephrase reno