Grammar mistake in 07_pegasos_qsvc.ipynb
In 07_pegasos_qsvc.ipynb, there's a sentence "This algorithm is ... beneficial of the kernel trick, and yields a training complexity...", but I didn't find any sentence examples using "beneficial of", I suppose it means "benefit from" here. It's a minor grammar mistake, but since it causes some misunderstandings, I think it would be better to fix it.